Death improvements
Now when a player is out of lives he's "out". If both players are out,
the game returns to the scene_intro, but if a player isnt out, he can
still play.
Also, now at the beginning of the game the players spawn as they should
(from the bottom of the screen)